+++ packages/core/src/index.ts
@@ -1,7 +1,6 @@
 export * from './config/config';
 export * from './route/route';
 export { Generator, GeneratorType } from './service/generator';
-export * from './service/env';
 export * from './service/pluginAPI';
 export * from './service/service';
 export { Env, IAdd, IEvent, IModify, IRoute } from './types';
