+++ packages/core/src/service/service.ts
@@ -242,6 +242,7 @@ export class Service {
     this.args = args;
     this.name = name;
 
+    // loadEnv
     this.stage = ServiceStage.init;
     loadEnv({ cwd: this.cwd, envFile: '.env' });
     // get pkg from package.json
