+++ packages/umi/src/client/utils.ts
@@ -1,5 +1,3 @@
-import { logger } from '@umijs/utils';
-
 export function assert(value: unknown, message: string) {
   if (!value) throw new Error(message);
 }
@@ -20,12 +18,4 @@ export function compose({
 
 export function isPromiseLike(obj: any) {
   return !!obj && typeof obj === 'object' && typeof obj.then === 'function';
-}
-
-export function printHelp() {
-  logger.fatal('A complete log of this run can be found in:');
-  logger.fatal(logger.getLatestLogFilePath());
-  logger.fatal(
-    'Consider reporting a GitHub issue : https://github.com/umijs/umi-next/issues',
-  );
 }
